+++ packages/adapter-next/tsup.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
     config: 'src/config.ts',
     client: 'src/client.ts',
     realtime: 'src/realtime.ts',
+    'validation-loader': 'src/validation-loader.ts',
     'realtime-definition-loader': 'src/realtime-definition-loader.ts',
     runtime: 'src/runtime.ts',
   },