+++ packages/validation/tsup.config.ts
@@ -5,6 +5,7 @@ const outDir = process.env.HOLO_BUILD_OUT_DIR ?? 'dist'
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
+    'internal/compiled': 'src/internal/compiled.ts',
   },
   format: ['esm'],
   dts: true,
@@ -13,5 +14,6 @@ export default defineConfig({
   outExtension: () => ({ js: '.mjs' }),
   esbuildOptions(options) {
     options.logLevel = 'warning'
+    options.pure = ['Object.freeze']
   },
 })